+++ utils/utilities.py
@@ -34,8 +34,9 @@ def make_model(num_layers,num_features,num_hidden_features,device):
     transform = CompositeTransform(transforms)
     
     #Uncomment the below if float64
-    #flow = Flow(transform, base_dist).double().to(device)
-    flow = Flow(transform, base_dist).to(device)
+    flow = Flow(transform, base_dist).double().to(device)
+    #Uncomment the below if float32
+    #flow = Flow(transform, base_dist).to(device)
     
     optimizer = optim.Adam(flow.parameters())
